Create Chris_Hughes_finite_groups
Stuff from M1P2 about groups, including lagrange's theorem

A few sorries. Proving that disjoint factorization of permutations is proving quite difficult, because a bit of stuff about disjoint sets is not in mathlib, although I have a PR about disjoint finsets prepared for when mathlib build passes for the latests lean nightly. `

`card_quot` was a particularly hard theorem `⟨_⟩` was the bit that took a very long time.